+++ crates/mongodb-agent-common/src/query/arguments.rs
@@ -3,12 +3,15 @@ use std::collections::BTreeMap;
 use indent::indent_all_by;
 use itertools::Itertools as _;
 use mongodb::bson::Bson;
-use serde_json::Value;
+use ndc_models::Argument;
 use thiserror::Error;
 
 use crate::mongo_query_plan::Type;
 
-use super::serialization::{json_to_bson, JsonToBsonError};
+use super::{
+    query_variable_name::query_variable_name,
+    serialization::{json_to_bson, JsonToBsonError},
+};
 
 #[derive(Debug, Error)]
 pub enum ArgumentError {
@@ -28,11 +31,11 @@ pub enum ArgumentError {
 /// map to declared parameters (no excess arguments).
 pub fn resolve_arguments(
     parameters: &BTreeMap<String, Type>,
-    mut arguments: BTreeMap<String, Value>,
+    mut arguments: BTreeMap<String, Argument>,
 ) -> Result<BTreeMap<String, Bson>, ArgumentError> {
     validate_no_excess_arguments(parameters, &arguments)?;
 
-    let (arguments, missing): (Vec<(String, Value, &Type)>, Vec<String>) = parameters
+    let (arguments, missing): (Vec<(String, Argument, &Type)>, Vec<String>) = parameters
         .iter()
         .map(|(name, parameter_type)| {
             if let Some((name, argument)) = arguments.remove_entry(name) {
@@ -48,12 +51,12 @@ pub fn resolve_arguments(
 
     let (resolved, errors): (BTreeMap<String, Bson>, BTreeMap<String, JsonToBsonError>) = arguments
         .into_iter()
-        .map(
-            |(name, argument, parameter_type)| match json_to_bson(parameter_type, argument) {
+        .map(|(name, argument, parameter_type)| {
+            match argument_to_mongodb_expression(&argument, parameter_type) {
                 Ok(bson) => Ok((name, bson)),
                 Err(err) => Err((name, err)),
-            },
-        )
+            }
+        })
         .partition_result();
     if !errors.is_empty() {
         return Err(ArgumentError::Invalid(errors));
@@ -62,9 +65,22 @@ pub fn resolve_arguments(
     Ok(resolved)
 }
 
-pub fn validate_no_excess_arguments(
+fn argument_to_mongodb_expression(
+    argument: &Argument,
+    parameter_type: &Type,
+) -> Result<Bson, JsonToBsonError> {
+    match argument {
+        Argument::Variable { name } => {
+            let mongodb_var_name = query_variable_name(name, parameter_type);
+            Ok(format!("$${mongodb_var_name}").into())
+        }
+        Argument::Literal { value } => json_to_bson(parameter_type, value.clone()),
+    }
+}
+
+pub fn validate_no_excess_arguments<T>(
     parameters: &BTreeMap<String, Type>,
-    arguments: &BTreeMap<String, Value>,
+    arguments: &BTreeMap<String, T>,
 ) -> Result<(), ArgumentError> {
     let excess: Vec<String> = arguments
         .iter()