+++ crates/test-helpers/src/lib.rs
@@ -1,5 +1,7 @@
 pub mod arb_bson;
+mod arb_plan_type;
 pub mod arb_type;
 
 pub use arb_bson::{arb_bson, arb_bson_with_options, ArbBsonOptions};
+pub use arb_plan_type::arb_plan_type;
 pub use arb_type::arb_type;