+++ fixtures/mongodb/sample_mflix/indexes.js
@@ -1,0 +1,3 @@
+db.comments.createIndex({ movie_id: 1 })
+db.comments.createIndex({ email: 1 })
+db.users.createIndex({ email: 1 })